requirements.txt
absl-py==2.0.0
accelerate==0.19.0
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.2.0
backcall==0.2.0
bitsandbytes==0.39.0
brotlipy==0.7.0
cachetools==5.3.2
cchardet==2.1.7
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
chardet==5.2.0
charset-normalizer==3.3.2
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
datasets==2.12.0
decorator==5.1.1
diffusers==0.21.0
dill==0.3.6
et-xmlfile==1.1.0
filelock==3.12.2
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
google-auth==2.25.2
google-auth-oauthlib==0.4.6
grpcio==1.60.0
huggingface-hub==0.16.4
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
inplace-abn==1.1.0
ipython==7.34.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
multidict==6.0.5
multiprocess==0.70.14
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==1.3.5
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.17.2
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.3.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tokenizers==0.13.3
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.66.1
traitlets==5.9.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==2.2.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.15.0
